Update Hero and Footer
