+++ components/sections/community/index.js
@@ -1,4 +1,4 @@
-import Section from '../../section'
+import Section from '../section'
 import Quotes from './quotes'
 import Logos from './logos'
