+++ pages/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Meta from '../components/Meta'
+import Meta from '../components/meta'
 
 // Sections
 import Hero from '../components/sections/hero'
